Simplifies unsafe data structures

+++ src/DotNext/ValueType.cs
@@ -74,35 +74,35 @@ namespace DotNext
             }
         }
 
-		/// <summary>
-		/// Checks bitwise equality between two values of different value types.
-		/// </summary>
-		/// <remarks>
-		/// This method doesn't use <see cref="object.Equals(object)"/>
-		/// even if it is overridden by value type.
-		/// </remarks>
-		/// <typeparam name="U">Type of second value.</typeparam>
-		/// <param name="first">The first value to check.</param>
-		/// <param name="second">The second value to check.</param>
-		/// <returns><see langword="true"/>, if both values are equal; otherwise, <see langword="false"/>.</returns>
-		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+        /// <summary>
+        /// Checks bitwise equality between two values of different value types.
+        /// </summary>
+        /// <remarks>
+        /// This method doesn't use <see cref="object.Equals(object)"/>
+        /// even if it is overridden by value type.
+        /// </remarks>
+        /// <typeparam name="U">Type of second value.</typeparam>
+        /// <param name="first">The first value to check.</param>
+        /// <param name="second">The second value to check.</param>
+        /// <returns><see langword="true"/>, if both values are equal; otherwise, <see langword="false"/>.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static unsafe bool BitwiseEquals<U>(T first, U second)
-			where U: struct
-            => Size == ValueType<U>.Size && Memory.Equals(Unsafe.AsPointer(ref first), Unsafe.AsPointer(ref second), Size);
+            where U : struct
+            => Size == ValueType<U>.Size && Memory.Equals(Unsafe.AsPointer(ref first), Unsafe.AsPointer(ref second), (long)Size);
 
-		/// <summary>
-		/// Checks bitwise equality between two values of the same value type.
-		/// </summary>
-		/// <remarks>
-		/// This method doesn't use <see cref="object.Equals(object)"/>
-		/// even if it is overridden by value type.
-		/// </remarks>
-		/// <param name="first">The first value to check.</param>
-		/// <param name="second">The second value to check.</param>
-		/// <returns><see langword="true"/>, if both values are equal; otherwise, <see langword="false"/>.</returns>
-		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+        /// <summary>
+        /// Checks bitwise equality between two values of the same value type.
+        /// </summary>
+        /// <remarks>
+        /// This method doesn't use <see cref="object.Equals(object)"/>
+        /// even if it is overridden by value type.
+        /// </remarks>
+        /// <param name="first">The first value to check.</param>
+        /// <param name="second">The second value to check.</param>
+        /// <returns><see langword="true"/>, if both values are equal; otherwise, <see langword="false"/>.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static unsafe bool BitwiseEquals(T first, T second)
-            => Memory.Equals(Unsafe.AsPointer(ref first), Unsafe.AsPointer(ref second), Size);
+            => Memory.Equals(Unsafe.AsPointer(ref first), Unsafe.AsPointer(ref second), (long)Size);
 
 		/// <summary>
 		/// Computes bitwise hash code for the specified value.
@@ -152,7 +152,7 @@ namespace DotNext
         /// <param name="second">The second value to compare.</param>
         /// <returns>A value that indicates the relative order of the objects being compared.</returns>
         public static unsafe int BitwiseCompare(T first, T second)
-            => Memory.Compare(Unsafe.AsPointer(ref first), Unsafe.AsPointer(ref second), Size);
+            => Memory.Compare(Unsafe.AsPointer(ref first), Unsafe.AsPointer(ref second), (long)Size);
 
 		/// <summary>
 		/// Compares bits of two values of the different type.
@@ -164,7 +164,7 @@ namespace DotNext
 		public static unsafe int BitwiseCompare<U>(T first, U second)
             where U: struct
             => Size == ValueType<U>.Size ? 
-					Memory.Compare(Unsafe.AsPointer(ref first), Unsafe.AsPointer(ref second), Size) :
+					Memory.Compare(Unsafe.AsPointer(ref first), Unsafe.AsPointer(ref second), (long)Size) :
 					Size.CompareTo(ValueType<U>.Size);
 
         /// <summary>
